third_party/blink/web_tests/platform/generic/fast/forms/setCustomValidity-arguments-expected.txt

Test how setCustomValidity reacts to too few arguments.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS b.setCustomValidity() threw exception TypeError: Failed to execute 'setCustomValidity' on 'HTMLButtonElement': 1 argument required, but only 0 present..
PASS fs.setCustomValidity() threw exception TypeError: Failed to execute 'setCustomValidity' on 'HTMLFieldSetElement': 1 argument required, but only 0 present..
PASS i.setCustomValidity() threw exception TypeError: Failed to execute 'setCustomValidity' on 'HTMLInputElement': 1 argument required, but only 0 present..
PASS successfullyParsed is true

TEST COMPLETE
